# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server eshop server.",
        "title": "Eshop API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
...
        "/admin_api/v1/property_values": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台规格值管理"
                ],
                "summary": "添加规格值",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.PropertyValueParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/property_values/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台规格值管理"
                ],
                "summary": "删除规格值",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "property_value id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
        "/admin_api/v1/sessions/forget_password": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理员"
                ],
                "summary": "管理员忘记密码",
                "parameters": [
                    {
                        "description": "重置密码",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ForgetPasswordParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/sessions/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理员"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "邮箱密码",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/sessions/mine": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理员"
                ],
                "summary": "当前管理员",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/sessions/reset_password": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理员"
                ],
                "summary": "管理员设置密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "充值密码token",
                        "name": "reset_password_token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "重置密码",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ResetPasswordParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
        "/admin_api/v1/web_pages": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台webview管理"
                ],
                "summary": "webview列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台webview管理"
                ],
                "summary": "添加webview",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.WebPageParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/web_pages/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台webview管理"
                ],
                "summary": "webview详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "web_page id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台webview管理"
                ],
                "summary": "更新webview",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "web_page params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.WebPageParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台webview管理"
                ],
                "summary": "删除webview",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "web_page id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiHelpers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "v1.ForgetPasswordParams": {
            "type": "object",
            "required": [
                "captcha_id",
                "captcha_value",
                "email"
            ],
            "properties": {
                "captcha_id": {
                    "type": "string"
                },
                "captcha_value": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
        "v1.LoginParams": {
            "type": "object",
            "required": [
                "captcha_id",
                "captcha_value",
                "email",
                "password"
            ],
            "properties": {
                "captcha_id": {
                    "type": "string"
                },
                "captcha_value": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "v1.PropertyValueParams": {
            "type": "object",
            "required": [
                "property_name_id",
                "value"
            ],
            "properties": {
                "_destroy": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "property_name_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "v1.ResetPasswordParams": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
...
        "v1.WebPageParams": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}